% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codebuild_operations.R
\name{batch_delete_builds}
\alias{batch_delete_builds}
\title{Deletes one or more builds}
\usage{
batch_delete_builds(ids)
}
\arguments{
\item{ids}{[required] The IDs of the builds to delete.}
}
\description{
Deletes one or more builds.
}
\section{Accepted Parameters}{
\preformatted{batch_delete_builds(
  ids = list(
    "string"
  )
)
}
}

