% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codebuild_operations.R
\name{batch_get_builds}
\alias{batch_get_builds}
\title{Gets information about builds}
\usage{
batch_get_builds(ids)
}
\arguments{
\item{ids}{The IDs of the builds.}
}
\description{
Gets information about builds.
}
\examples{
# The following example gets information about builds with the specified
# build IDs.
batch_get_builds(
  ids = list(
    "codebuild-demo-project:9b0ac37f-d19e-4254-9079-f47e9a389eEX",
    "codebuild-demo-project:b79a46f7-1473-4636-a23f-da9c45c208EX"
  )
)

}
