% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codebuild_operations.R
\name{batch_get_projects}
\alias{batch_get_projects}
\title{Gets information about build projects}
\usage{
batch_get_projects(names)
}
\arguments{
\item{names}{[required] The names of the build projects.}
}
\description{
Gets information about build projects.
}
\section{Accepted Parameters}{
\preformatted{batch_get_projects(
  names = list(
    "string"
  )
)
}
}

