% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codebuild_operations.R
\name{delete_project}
\alias{delete_project}
\title{Deletes a build project}
\usage{
delete_project(
  name = "string"
)
}
\arguments{
\item{name}{[required] The name of the build project.}
}
\description{
Deletes a build project.
}
\examples{

}
