% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codebuild_operations.R
\name{delete_source_credentials}
\alias{delete_source_credentials}
\title{Deletes a set of GitHub, GitHub Enterprise, or Bitbucket source credentials}
\usage{
delete_source_credentials(arn)
}
\arguments{
\item{arn}{[required] The Amazon Resource Name (ARN) of the token.}
}
\description{
Deletes a set of GitHub, GitHub Enterprise, or Bitbucket source credentials.
}
\section{Accepted Parameters}{
\preformatted{delete_source_credentials(
  arn = "string"
)
}
}

