% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codebuild_operations.R
\name{delete_webhook}
\alias{delete_webhook}
\title{For an existing AWS CodeBuild build project that has its source code stored in a GitHub repository, stops AWS CodeBuild from automatically rebuilding the source code every time a code change is pushed to the repository}
\usage{
delete_webhook(projectName)
}
\arguments{
\item{projectName}{The name of the AWS CodeBuild project.}
}
\description{
For an existing AWS CodeBuild build project that has its source code stored in a GitHub repository, stops AWS CodeBuild from automatically rebuilding the source code every time a code change is pushed to the repository.
}
\examples{

}
