% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codebuild_operations.R
\name{import_source_credentials}
\alias{import_source_credentials}
\title{Imports the source repository credentials for an AWS CodeBuild project that has its source code stored in a GitHub, GitHub Enterprise, or Bitbucket repository}
\usage{
import_source_credentials(username = NULL, token, serverType, authType)
}
\arguments{
\item{username}{The Bitbucket username when the \code{authType} is BASIC\_AUTH. This parameter is not valid for other types of source providers or connections.}

\item{token}{[required] For GitHub or GitHub Enterprise, this is the personal access token. For Bitbucket, this is the app password.}

\item{serverType}{[required] The source provider used for this project.}

\item{authType}{[required] The type of authentication used to connect to a GitHub, GitHub Enterprise, or Bitbucket repository. An OAUTH connection is not supported by the API and must be created using the AWS CodeBuild console.}
}
\description{
Imports the source repository credentials for an AWS CodeBuild project that has its source code stored in a GitHub, GitHub Enterprise, or Bitbucket repository.
}
\section{Accepted Parameters}{
\preformatted{import_source_credentials(
  username = "string",
  token = "string",
  serverType = "GITHUB"|"BITBUCKET"|"GITHUB_ENTERPRISE",
  authType = "OAUTH"|"BASIC_AUTH"|"PERSONAL_ACCESS_TOKEN"
)
}
}

