% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codebuild_operations.R
\name{invalidate_project_cache}
\alias{invalidate_project_cache}
\title{Resets the cache for a project}
\usage{
invalidate_project_cache(projectName)
}
\arguments{
\item{projectName}{[required] The name of the AWS CodeBuild build project that the cache will be reset for.}
}
\description{
Resets the cache for a project.
}
\section{Accepted Parameters}{
\preformatted{invalidate_project_cache(
  projectName = "string"
)
}
}

