% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codebuild_operations.R
\name{list_curated_environment_images}
\alias{list_curated_environment_images}
\title{Gets information about Docker images that are managed by AWS CodeBuild}
\usage{
list_curated_environment_images()
}
\description{
Gets information about Docker images that are managed by AWS CodeBuild.
}
\section{Accepted Parameters}{
\preformatted{list_curated_environment_images()
}
}

