% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codebuild_operations.R
\name{list_source_credentials}
\alias{list_source_credentials}
\title{Returns a list of SourceCredentialsInfo objects}
\usage{
list_source_credentials()
}
\description{
Returns a list of \code{SourceCredentialsInfo} objects.
}
\section{Accepted Parameters}{
\preformatted{list_source_credentials()
}
}

