% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codebuild_package.R
\docType{package}
\name{paws.codebuild-package}
\alias{paws.codebuild}
\alias{paws.codebuild-package}
\title{paws.codebuild: AWS CodeBuild}
\description{
AWS CodeBuild is a fully managed build service in
    the cloud. AWS CodeBuild compiles your source code, runs unit tests,
    and produces artifacts that are ready to deploy. AWS CodeBuild
    eliminates the need to provision, manage, and scale your own build
    servers. It provides prepackaged build environments for the most
    popular programming languages and build tools, such as Apache Maven,
    Gradle, and more. You can also fully customize build environments in
    AWS CodeBuild to use your own build tools. AWS CodeBuild scales
    automatically to meet peak build requests, and you pay only for the
    build time you consume. For more information about AWS CodeBuild, see
    the *AWS CodeBuild User Guide*.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
