% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codebuild_operations.R
\name{stop_build}
\alias{stop_build}
\title{Attempts to stop running a build}
\usage{
stop_build(
  id = "string"
)
}
\arguments{
\item{id}{[required] The ID of the build.}
}
\description{
Attempts to stop running a build.
}
\examples{

}
