% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codebuild_operations.R
\name{update_project}
\alias{update_project}
\title{Changes the settings of a build project}
\usage{
update_project(name, description = NULL, source = NULL,
  secondarySources = NULL, artifacts = NULL, secondaryArtifacts = NULL,
  cache = NULL, environment = NULL, serviceRole = NULL,
  timeoutInMinutes = NULL, encryptionKey = NULL, tags = NULL,
  vpcConfig = NULL, badgeEnabled = NULL, logsConfig = NULL)
}
\arguments{
\item{name}{The name of the build project.

You cannot change a build project's name.}

\item{description}{A new or replacement description of the build project.}

\item{source}{Information to be changed about the build input source code for the build project.}

\item{secondarySources}{An array of \code{ProjectSource} objects.}

\item{artifacts}{Information to be changed about the build output artifacts for the build project.}

\item{secondaryArtifacts}{An array of \code{ProjectSource} objects.}

\item{cache}{Stores recently used information so that it can be quickly accessed at a later time.}

\item{environment}{Information to be changed about the build environment for the build project.}

\item{serviceRole}{The replacement ARN of the AWS Identity and Access Management (IAM) role that enables AWS CodeBuild to interact with dependent AWS services on behalf of the AWS account.}

\item{timeoutInMinutes}{The replacement value in minutes, from 5 to 480 (8 hours), for AWS CodeBuild to wait before timing out any related build that did not get marked as completed.}

\item{encryptionKey}{The replacement AWS Key Management Service (AWS KMS) customer master key (CMK) to be used for encrypting the build output artifacts.

You can specify either the CMK's Amazon Resource Name (ARN) or, if available, the CMK's alias (using the format \code{alias/alias-name alias-name }).}

\item{tags}{The replacement set of tags for this build project.

These tags are available for use by AWS services that support AWS CodeBuild build project tags.}

\item{vpcConfig}{VpcConfig enables AWS CodeBuild to access resources in an Amazon VPC.}

\item{badgeEnabled}{Set this to true to generate a publicly-accessible URL for your project's build badge.}

\item{logsConfig}{Information about logs for the build project. A project can create Amazon CloudWatch Logs, logs in an S3 bucket, or both.}
}
\description{
Changes the settings of a build project.
}
\examples{

}
