% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codebuild_operations.R
\name{update_webhook}
\alias{update_webhook}
\title{Updates the webhook associated with an AWS CodeBuild build project}
\usage{
update_webhook(projectName, branchFilter = NULL, rotateSecret = NULL)
}
\arguments{
\item{projectName}{The name of the AWS CodeBuild project.}

\item{branchFilter}{A regular expression used to determine which branches in a repository are built when a webhook is triggered. If the name of a branch matches the regular expression, then it is built. If it doesn't match, then it is not. If branchFilter is empty, then all branches are built.}

\item{rotateSecret}{A boolean value that specifies whether the associated repository's secret token should be updated.}
}
\description{
Updates the webhook associated with an AWS CodeBuild build project.
}
\examples{

}
