% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.codebuild_operations.R
\name{update_webhook}
\alias{update_webhook}
\title{Updates the webhook associated with an AWS CodeBuild build project}
\usage{
update_webhook(projectName, branchFilter = NULL, rotateSecret = NULL)
}
\arguments{
\item{projectName}{[required] The name of the AWS CodeBuild project.}

\item{branchFilter}{A regular expression used to determine which repository branches are built when a webhook is triggered. If the name of a branch matches the regular expression, then it is built. If \code{branchFilter} is empty, then all branches are built.}

\item{rotateSecret}{A boolean value that specifies whether the associated GitHub repository's secret token should be updated. If you use Bitbucket for your repository, \code{rotateSecret} is ignored.}
}
\description{
Updates the webhook associated with an AWS CodeBuild build project.
}
\details{
If you use Bitbucket for your repository, \code{rotateSecret} is ignored.
}
\section{Accepted Parameters}{
\preformatted{update_webhook(
  projectName = "string",
  branchFilter = "string",
  rotateSecret = TRUE|FALSE
)
}
}

